Extract extracting jar from resources to JarUtils
